refactor header logic
