+++ frontend/src/app/layout.tsx
@@ -2,7 +2,7 @@ import { Analytics } from "@vercel/analytics/next";
 import type { Metadata } from "next";
 import { Modak, Nunito } from "next/font/google";
 
-import Header from "@/components/header/header";
+import Header from "@/features/header/components/header";
 import { Providers } from "@/lib/providers";
 import "@/styles/globals.css";
