+++ frontend/src/features/header/components/buttons/account.tsx
@@ -4,11 +4,11 @@ import { LogOutIcon, UserIcon } from "lucide-react";
 import { useRouter } from "next/navigation";
 
 import KebabMenu from "@/components/kebab-menu";
-import { getSession } from "@/features/account/get-session";
+import { AccountDetails } from "@/features/account/type";
 import ActionButton from "@/features/button/components/action";
 import EmptyButton from "@/features/button/components/empty";
 import LinkButton from "@/features/button/components/link";
-import ShrinkingHeaderButton from "@/features/header/components/shrinking-header-button";
+import ShrinkingHeaderButton from "@/features/header/components/buttons/shrinking-header";
 import { useHeaderSize } from "@/features/header/context";
 import { useToast } from "@/features/system-feedback";
 import { MESSAGES } from "@/lib/messages";
@@ -16,9 +16,12 @@ import { clientPost } from "@/lib/utils/api/client-fetch";
 import { ROUTES } from "@/lib/utils/api/endpoints";
 import { ApiErrorResponse } from "@/lib/utils/api/fetch-wrapper";
 
-export default function AccountButton() {
+export default function AccountButton({
+  accountDetails,
+}: {
+  accountDetails: AccountDetails;
+}) {
   const { activeMenu, setActiveMenu } = useHeaderSize();
-  const { loginState, logout, accountDetails } = useAccount();
   const router = useRouter();
   const { addToast } = useToast();
 
@@ -27,13 +30,14 @@ export default function AccountButton() {
   const signOut = async () => {
     try {
       await clientPost(ROUTES.auth.logout);
-      redirect("/login");
-      // cannot use toast after redirect bc toast does not work on server components
+      addToast("success", MESSAGES.SUCCESS_LOGOUT);
+
+      router.push("/login");
+      router.refresh();
     } catch (e) {
       const error = e as ApiErrorResponse;
       console.error("Logout error:", error);
-      // addToast("error", error.formattedMessage);
-      return false;
+      addToast("error", error.formattedMessage);
     }
   };
 
@@ -55,43 +59,30 @@ export default function AccountButton() {
     />
   );
 
-  if (accountDetails) {
-    return (
-      <ShrinkingHeaderButton
-        buttonStyle="frosted glass inset"
-        icon={<UserIcon />}
-      >
-        <KebabMenu
-          nested
-          open={isMenuOpen}
-          onOpenChange={(isOpen) => setActiveMenu(isOpen ? "account" : null)}
-          trigger={
-            <EmptyButton
-              className="relative z-10"
-              buttonStyle="frosted glass inset"
-              icon={<UserIcon />}
-              aria-label="Account settings"
-            />
-          }
-        >
-          <h2 className="text-foreground text-center font-bold">
-            {accountDetails?.email}
-          </h2>
-          {accountSettingsButton}
-          {signOutButton}
-        </KebabMenu>
-      </ShrinkingHeaderButton>
-    );
-  }
-
   return (
-    <ShrinkingHeaderButton buttonStyle="frosted glass inset" label="Log In">
-      <LinkButton
-        className="relative z-10"
-        buttonStyle="frosted glass inset"
-        label="Log In"
-        href="/login"
-      />
+    <ShrinkingHeaderButton
+      buttonStyle="frosted glass inset"
+      icon={<UserIcon />}
+    >
+      <KebabMenu
+        nested
+        open={isMenuOpen}
+        onOpenChange={(isOpen) => setActiveMenu(isOpen ? "account" : null)}
+        trigger={
+          <EmptyButton
+            className="relative z-10"
+            buttonStyle="frosted glass inset"
+            icon={<UserIcon />}
+            aria-label="Account settings"
+          />
+        }
+      >
+        <h2 className="text-foreground text-center font-bold">
+          {accountDetails?.email}
+        </h2>
+        {accountSettingsButton}
+        {signOutButton}
+      </KebabMenu>
     </ShrinkingHeaderButton>
   );
 }
