+++ frontend/src/features/header/components/buttons/new-event.tsx
@@ -1,11 +1,9 @@
-"use client";
-
 import { PlusIcon } from "lucide-react";
 import { usePathname } from "next/navigation";
 
 import LinkButton from "@/features/button/components/link";
 import { ButtonStyle } from "@/features/button/props";
-import ShrinkingHeaderButton from "@/features/header/components/shrinking-header-button";
+import ShrinkingHeaderButton from "@/features/header/components/buttons/shrinking-header";
 
 export default function NewEventButton() {
   const pathname = usePathname();
