+++ frontend/src/features/header/components/buttons/shrinking-header.tsx
