+++ frontend/src/features/header/components/shrinking-header.tsx
@@ -1,0 +1,124 @@
+"use client";
+
+import { useEffect, useRef, useState } from "react";
+
+import { motion } from "framer-motion";
+
+import DashboardButton from "@/features/header/components/buttons/dashboard";
+import NewEventButton from "@/features/header/components/buttons/new-event";
+import ThemeToggle from "@/features/header/components/buttons/theme-toggle";
+import LogoArea from "@/features/header/components/logo-area";
+import { useHeaderSize } from "@/features/header/context";
+import useCheckMobile from "@/lib/hooks/use-check-mobile";
+import { cn } from "@/lib/utils/classname";
+
+const SCROLL_THRESHOLD = 50;
+
+export default function ShrinkingHeader({
+  children: accountButton,
+}: {
+  children: React.ReactNode;
+}) {
+  const [mounted, setMounted] = useState(false);
+
+  const isMobile = useCheckMobile();
+  const lastScrollPoint = useRef(0);
+  const scrollCheckpoint = useRef(0);
+
+  const { isShrunk, heightClass, shrink, expand, activeMenu } = useHeaderSize();
+
+  useEffect(() => {
+    setMounted(true);
+
+    if (!isMobile) {
+      expand();
+      return;
+    }
+
+    const handleScroll = () => {
+      const currentScrollPoint = Math.min(
+        Math.max(window.scrollY, 0),
+        document.documentElement.scrollHeight - window.innerHeight,
+      );
+      const scrollingDown = currentScrollPoint > lastScrollPoint.current;
+      lastScrollPoint.current = currentScrollPoint;
+
+      if (currentScrollPoint <= 0) {
+        expand();
+        return;
+      }
+
+      if (isShrunk) {
+        if (scrollingDown) {
+          scrollCheckpoint.current = currentScrollPoint;
+        } else if (
+          currentScrollPoint <
+          scrollCheckpoint.current - SCROLL_THRESHOLD
+        ) {
+          expand();
+        }
+      } else {
+        if (!scrollingDown) {
+          scrollCheckpoint.current = currentScrollPoint;
+        } else if (
+          currentScrollPoint >
+          scrollCheckpoint.current + SCROLL_THRESHOLD
+        ) {
+          shrink();
+        }
+      }
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, [isMobile, isShrunk, shrink, expand]);
+
+  if (!mounted) return null;
+
+  return (
+    <header className={cn(heightClass, "fixed top-0 z-40 w-full pt-4")}>
+      <nav
+        className={cn(
+          "flex w-full max-w-[1440px] justify-between px-4",
+          isShrunk ? "cursor-pointer" : "",
+        )}
+        onClickCapture={(e) => {
+          if (!isShrunk) return;
+          e.preventDefault();
+          e.stopPropagation();
+          expand();
+        }}
+      >
+        <LogoArea isShrunk={isShrunk} />
+
+        <motion.div
+          layout
+          animate={{ scale: activeMenu ? 0.95 : 1 }}
+          transition={{ duration: 0.3, ease: "easeInOut" }}
+          className={cn(
+            "frosted-glass relative flex h-fit items-center rounded-full",
+            "header-transition-[gap,padding]",
+            isShrunk ? "gap-1 p-1" : "gap-2 p-2",
+          )}
+        >
+          <NewEventButton />
+          <ThemeToggle />
+          <DashboardButton />
+
+          {accountButton}
+
+          <div
+            className={cn(
+              "bg-violet/20 pointer-events-none absolute inset-0 rounded-full",
+              "transition-opacity duration-300 ease-in-out",
+              activeMenu ? "opacity-100" : "opacity-0",
+            )}
+            aria-hidden="true"
+          />
+        </motion.div>
+      </nav>
+    </header>
+  );
+}
